+++ lib/rspec/matchers/operator_matcher.rb
@@ -60,11 +60,13 @@ module RSpec
         "#{@operator} #{@expected.inspect}"
       end
 
-      private
+    private
 
       if Method.method_defined?(:owner) # 1.8.6 lacks Method#owner :-(
         def uses_generic_implementation_of?(op)
-          @actual.method(op).owner == ::Kernel
+          Expectations.method_handle_for(@actual, op).owner == ::Kernel
+        rescue NameError
+          false
         end
       else
         def uses_generic_implementation_of?(op)
@@ -74,7 +76,9 @@ module RSpec
           #
           # In the absence of Method#owner, this is the best we
           # can do to see if the method comes from Kernel.
-          @actual.method(op).to_s.include?('(Kernel)')
+          Expectations.method_handle_for(@actual, op).to_s.include?('(Kernel)')
+        rescue NameError
+          false
         end
       end