+++ spec/rspec/matchers/operator_matcher_spec.rb
@@ -1,5 +1,19 @@
 require 'spec_helper'
 
+class MethodOverrideObject
+  def method
+    :foo
+  end
+end
+
+class MethodMissingObject < Struct.new(:original)
+  undef ==
+
+  def method_missing(name, *args, &block)
+    original.__send__ name, *args, &block
+  end
+end
+
 describe "operator matchers", :uses_should do
   describe "should ==" do
     it "delegates message to target" do
@@ -17,6 +31,21 @@ describe "operator matchers", :uses_should do
       subject = "apple"
       RSpec::Expectations.should_receive(:fail_with).with(%[expected: "orange"\n     got: "apple" (using ==)], "orange", "apple")
       subject.should == "orange"
+    end
+
+    it "works when #method is overriden" do
+      myobj = MethodOverrideObject.new
+      expect {
+        myobj.should == myobj
+      }.to_not raise_error
+    end
+
+    it "works when implemented via method_missing" do
+      obj = Object.new
+
+      myobj = MethodMissingObject.new(obj)
+      myobj.should == obj
+      myobj.should_not == Object.new
     end
   end