+++ Changelog.md
@@ -1,3 +1,9 @@
+### development
+
+Bug fixes:
+
+* Make the `match` matcher produce a diff output (Jon Rowe, Ben Moss)
+
 ### 2.14.3 / 2013-09-22
 [full changelog](http://github.com/rspec/rspec-expectations/compare/v2.14.2...v2.14.3)